Changed the type of the ram size to unsigned long. Ram size cannot be negative and we want to use the full range of allocatable memory on a 64bit host.

+++ cpu-all.h
@@ -880,7 +880,7 @@ int argos_cpu_inl(CPUState *env, int addr, argos_rtag_t *tag);
 
 /* memory API */
 
-extern int phys_ram_size;
+extern unsigned long phys_ram_size;
 extern int phys_ram_fd;
 extern uint8_t *phys_ram_base;
 extern uint8_t *phys_ram_dirty;
